Fix keyboard and mosue queue events

+++ src/kernel/devices/mouse.h
@@ -13,9 +13,9 @@
 
 struct mouse_event
 {
-	uint32_t x;
-	uint32_t y;
-	uint32_t buttons;
+	int32_t x;
+	int32_t y;
+	uint8_t buttons;
 } mouse_event;
 
 struct mouse_inode
